+++ src/test/groovy/org/prebid/server/functional/tests/privacy/CcpaAuctionSpec.groovy
@@ -1,24 +1,19 @@
 package org.prebid.server.functional.tests.privacy
 
 import org.prebid.server.functional.model.ChannelType
-import org.prebid.server.functional.model.bidder.BidderName
 import org.prebid.server.functional.model.config.AccountCcpaConfig
 import org.prebid.server.functional.model.config.AccountConfig
 import org.prebid.server.functional.model.config.AccountPrivacyConfig
 import org.prebid.server.functional.model.db.Account
-import org.prebid.server.functional.model.db.StoredRequest
-import org.prebid.server.functional.model.request.auction.BidRequest
 import org.prebid.server.functional.model.request.auction.DistributionChannel
-import org.prebid.server.functional.model.response.auction.ErrorType
 import org.prebid.server.functional.testcontainers.PBSTest
 import org.prebid.server.functional.util.privacy.BogusConsent
 import org.prebid.server.functional.util.privacy.CcpaConsent
-import org.prebid.server.functional.util.privacy.TcfConsent
 import spock.lang.PendingFeature
 
-import static org.prebid.server.functional.model.request.amp.ConsentType.BOGUS
+import static org.prebid.server.functional.model.ChannelType.WEB
+import static org.prebid.server.functional.model.bidder.BidderName.GENERIC
 import static org.prebid.server.functional.util.privacy.CcpaConsent.Signal.ENFORCED
-import static org.prebid.server.functional.util.privacy.TcfConsent.PurposeId.BASIC_ADS
 
 @PBSTest
 class CcpaAuctionSpec extends PrivacyBaseSpec {
@@ -30,11 +25,8 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         def bidRequest = getCcpaBidRequest(validCcpa)
 
         and: "Save account config into DB"
-        def ccpa = new AccountCcpaConfig(enabled: true)
-        def privacy = new AccountPrivacyConfig(ccpa: ccpa)
-        def accountConfig = new AccountConfig(privacy: privacy)
-        def account = new Account(uuid: bidRequest.site.publisher.id, config: accountConfig)
-        accountDao.save(account)
+        def ccpaConfig = new AccountCcpaConfig(enabled: true)
+        accountDao.save(getAccountWithCcpa(bidRequest.site.publisher.id, ccpaConfig))
 
         when: "PBS processes auction request"
         defaultPbsService.sendAuctionRequest(bidRequest)
@@ -51,11 +43,8 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         def bidRequest = getCcpaBidRequest(validCcpa)
 
         and: "Save account config into DB"
-        def ccpa = new AccountCcpaConfig(enabled: false)
-        def privacy = new AccountPrivacyConfig(ccpa: ccpa)
-        def accountConfig = new AccountConfig(privacy: privacy)
-        def account = new Account(uuid: bidRequest.site.publisher.id, config: accountConfig)
-        accountDao.save(account)
+        def ccpaConfig = new AccountCcpaConfig(enabled: false)
+        accountDao.save(getAccountWithCcpa(bidRequest.site.publisher.id, ccpaConfig))
 
         when: "PBS processes auction request"
         defaultPbsService.sendAuctionRequest(bidRequest)
@@ -95,7 +84,7 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
             !privacy.resolvedPrivacy?.tcf?.tcfConsentVersion
             !privacy.resolvedPrivacy?.tcf?.inEea
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC] ==
+            privacy.privacyActionsPerBidder[GENERIC] ==
                     ["Geolocation was masked in request to bidder according to CCPA policy."]
 
             privacy.errors?.isEmpty()
@@ -122,154 +111,11 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
             privacy.originPrivacy?.ccpa?.usPrivacy == invalidCcpa as String
             privacy.resolvedPrivacy?.ccpa?.usPrivacy == invalidCcpa as String
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC].isEmpty()
+            privacy.privacyActionsPerBidder[GENERIC].isEmpty()
 
             privacy.errors == ["CCPA consent $invalidCcpa has invalid format: us_privacy must specify 'N' " +
                                        "or 'n', 'Y' or 'y', '-' for the explicit notice" as String]
         }
-    }
-
-    @PendingFeature
-    def "PBS should add debug log for amp request when valid ccpa was passed"() {
-        given: "Default AmpRequest"
-        def validCcpa = new CcpaConsent(explicitNotice: ENFORCED, optOutSale: ENFORCED)
-        def ampRequest = getCcpaAmpRequest(validCcpa)
-
-        and: "Save storedRequest into DB"
-        def ampStoredRequest = bidRequestWithGeo
-        def storedRequest = StoredRequest.getDbStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        def response = defaultPbsService.sendAmpRequest(ampRequest)
-
-        then: "Response should contain debug log"
-        assert response.ext?.debug?.privacy
-        def privacy = response.ext?.debug?.privacy
-
-        verifyAll {
-            privacy.originPrivacy?.ccpa?.usPrivacy == validCcpa as String
-            privacy.resolvedPrivacy?.ccpa?.usPrivacy == validCcpa as String
-
-            !privacy.originPrivacy?.coppa?.coppa
-            !privacy.resolvedPrivacy?.coppa?.coppa
-
-            !privacy.originPrivacy?.tcf?.gdpr
-            !privacy.originPrivacy?.tcf?.tcfConsentString
-            !privacy.originPrivacy?.tcf?.tcfConsentVersion
-            !privacy.originPrivacy?.tcf?.inEea
-            !privacy.resolvedPrivacy?.tcf?.gdpr
-            !privacy.resolvedPrivacy?.tcf?.tcfConsentString
-            !privacy.resolvedPrivacy?.tcf?.tcfConsentVersion
-            !privacy.resolvedPrivacy?.tcf?.inEea
-
-            privacy.privacyActionsPerBidder[BidderName.GENERIC] ==
-                    ["Geolocation was masked in request to bidder according to CCPA policy."]
-
-            privacy.errors?.isEmpty()
-        }
-    }
-
-    @PendingFeature
-    def "PBS should add debug log for amp request when invalid ccpa was passed"() {
-        given: "Default AmpRequest"
-        def invalidCcpa = new BogusConsent()
-        def ampRequest = getCcpaAmpRequest(invalidCcpa)
-
-        and: "Save storedRequest into DB"
-        def ampStoredRequest = bidRequestWithGeo
-        def storedRequest = StoredRequest.getDbStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        def response = defaultPbsService.sendAmpRequest(ampRequest)
-
-        then: "Response should contain debug log with error"
-        assert response.ext?.debug?.privacy
-        def privacy = response.ext?.debug?.privacy
-
-        verifyAll {
-            privacy.originPrivacy?.ccpa?.usPrivacy == invalidCcpa as String
-            privacy.resolvedPrivacy?.ccpa?.usPrivacy == invalidCcpa as String
-
-            privacy.privacyActionsPerBidder[BidderName.GENERIC].isEmpty()
-
-            privacy.errors ==
-                    ["Amp request parameter consent_string or gdpr_consent have invalid format: $invalidCcpa" as String]
-        }
-    }
-
-    def "PBS should emit error for amp request when consent_string contains invalid ccpa consent"() {
-        given: "Default AmpRequest with invalid ccpa consent"
-        def ampRequest = getCcpaAmpRequest(invalidCcpa)
-        def ampStoredRequest = BidRequest.defaultBidRequest.tap {
-            site.publisher.id = ampRequest.account
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getDbStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        def response = defaultPbsService.sendAmpRequest(ampRequest)
-
-        then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message ==
-                ["Amp request parameter consent_string has invalid format for consent " +
-                         "type usPrivacy: $invalidCcpa" as String]
-
-        where:
-        invalidCcpa << [new BogusConsent(), new TcfConsent.Builder()
-                .setPurposesLITransparency(BASIC_ADS)
-                .build()]
-    }
-
-    def "PBS should emit error for amp request with gdprConsent when consent_type is invalid"() {
-        given: "Default AmpRequest with invalid ccpa type"
-        def validCcpa = new CcpaConsent(explicitNotice: ENFORCED, optOutSale: ENFORCED)
-        def ampRequest = getCcpaAmpRequest(validCcpa).tap {
-            consentType = BOGUS
-        }
-        def ampStoredRequest = BidRequest.defaultBidRequest.tap {
-            site.publisher.id = ampRequest.account
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getDbStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        def response = defaultPbsService.sendAmpRequest(ampRequest)
-
-        then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message == ["Invalid consent_type param passed"]
-    }
-
-    def "PBS should emit error for amp request when set not appropriate tcf consent"() {
-        given: "Default AmpRequest with gdpr_consent"
-        def tcfConsent = new TcfConsent.Builder()
-                .setPurposesLITransparency(BASIC_ADS)
-                .build()
-        def ampRequest = getCcpaAmpRequest(null).tap {
-            gdprConsent = tcfConsent
-        }
-        def ampStoredRequest = BidRequest.defaultBidRequest.tap {
-            site.publisher.id = ampRequest.account
-        }
-
-        and: "Save storedRequest into DB"
-        def storedRequest = StoredRequest.getDbStoredRequest(ampRequest, ampStoredRequest)
-        storedRequestDao.save(storedRequest)
-
-        when: "PBS processes amp request"
-        def response = defaultPbsService.sendAmpRequest(ampRequest)
-
-        then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message ==
-                ["Amp request parameter gdpr_consent has invalid format for consent type usPrivacy: $tcfConsent" as String]
     }
 
     def "PBS should apply ccpa when privacy.ccpa.channel-enabled.app or privacy.ccpa.enabled = true in account config"() {
@@ -278,10 +124,7 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         def bidRequest = getCcpaBidRequest(DistributionChannel.APP, validCcpa)
 
         and: "Save account config into DB"
-        def privacy = new AccountPrivacyConfig(ccpa: ccpaConfig)
-        def accountConfig = new AccountConfig(privacy: privacy)
-        def account = new Account(uuid: bidRequest.app.publisher.id, config: accountConfig)
-        accountDao.save(account)
+        accountDao.save(getAccountWithCcpa(bidRequest.app.publisher.id, ccpaConfig))
 
         when: "PBS processes auction request"
         defaultPbsService.sendAuctionRequest(bidRequest)
@@ -301,10 +144,7 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         def bidRequest = getCcpaBidRequest(validCcpa)
 
         and: "Save account config into DB"
-        def privacy = new AccountPrivacyConfig(ccpa: ccpaConfig)
-        def accountConfig = new AccountConfig(privacy: privacy)
-        def account = new Account(uuid: bidRequest.site.publisher.id, config: accountConfig)
-        accountDao.save(account)
+        accountDao.save(getAccountWithCcpa(bidRequest.site.publisher.id, ccpaConfig))
 
         when: "PBS processes auction request"
         defaultPbsService.sendAuctionRequest(bidRequest)
@@ -314,7 +154,7 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         assert bidderRequests.device?.geo == maskGeo(bidRequest)
 
         where:
-        ccpaConfig << [new AccountCcpaConfig(enabled: false, channelEnabled: [(ChannelType.WEB): true]),
+        ccpaConfig << [new AccountCcpaConfig(enabled: false, channelEnabled: [(WEB): true]),
                        new AccountCcpaConfig(enabled: true)]
     }
 
@@ -324,10 +164,7 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         def bidRequest = getCcpaBidRequest(DistributionChannel.APP, validCcpa)
 
         and: "Save account config into DB"
-        def privacy = new AccountPrivacyConfig(ccpa: ccpaConfig)
-        def accountConfig = new AccountConfig(privacy: privacy)
-        def account = new Account(uuid: bidRequest.app.publisher.id, config: accountConfig)
-        accountDao.save(account)
+        accountDao.save(getAccountWithCcpa(bidRequest.app.publisher.id, ccpaConfig))
 
         when: "PBS processes auction request"
         defaultPbsService.sendAuctionRequest(bidRequest)
@@ -348,10 +185,7 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         def bidRequest = getCcpaBidRequest(validCcpa)
 
         and: "Save account config into DB"
-        def privacy = new AccountPrivacyConfig(ccpa: ccpaConfig)
-        def accountConfig = new AccountConfig(privacy: privacy)
-        def account = new Account(uuid: bidRequest.site.publisher.id, config: accountConfig)
-        accountDao.save(account)
+        accountDao.save(getAccountWithCcpa(bidRequest.site.publisher.id, ccpaConfig))
 
         when: "PBS processes auction request"
         defaultPbsService.sendAuctionRequest(bidRequest)
@@ -362,7 +196,13 @@ class CcpaAuctionSpec extends PrivacyBaseSpec {
         assert bidderRequests.device?.geo?.lon == bidRequest.device.geo.lon
 
         where:
-        ccpaConfig << [new AccountCcpaConfig(enabled: true, channelEnabled: [(ChannelType.WEB): false]),
+        ccpaConfig << [new AccountCcpaConfig(enabled: true, channelEnabled: [(WEB): false]),
                        new AccountCcpaConfig(enabled: false)]
+    }
+
+    private Account getAccountWithCcpa(String accountId, AccountCcpaConfig ccpaConfig) {
+        def privacy = new AccountPrivacyConfig(ccpa: ccpaConfig)
+        def accountConfig = new AccountConfig(privacy: privacy)
+        new Account(uuid: accountId, config: accountConfig)
     }
 }
