+++ src/test/groovy/org/prebid/server/functional/tests/privacy/CoppaSpec.groovy
@@ -1,10 +1,11 @@
 package org.prebid.server.functional.tests.privacy
 
-import org.prebid.server.functional.model.bidder.BidderName
 import org.prebid.server.functional.model.db.StoredRequest
 import org.prebid.server.functional.model.request.amp.AmpRequest
 import org.prebid.server.functional.testcontainers.PBSTest
 import spock.lang.PendingFeature
+
+import static org.prebid.server.functional.model.bidder.BidderName.GENERIC
 
 @PBSTest
 class CoppaSpec extends PrivacyBaseSpec {
@@ -39,7 +40,7 @@ class CoppaSpec extends PrivacyBaseSpec {
             !privacy.originPrivacy?.ccpa?.usPrivacy
             !privacy.resolvedPrivacy?.ccpa?.usPrivacy
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC].isEmpty()
+            privacy.privacyActionsPerBidder[GENERIC].isEmpty()
 
             privacy.errors?.isEmpty()
         }
@@ -63,7 +64,7 @@ class CoppaSpec extends PrivacyBaseSpec {
             privacy.originPrivacy?.coppa?.coppa == bidRequest.regs.coppa
             privacy.resolvedPrivacy?.coppa?.coppa == bidRequest.regs.coppa
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC] ==
+            privacy.privacyActionsPerBidder[GENERIC] ==
                     ["Geolocation and address were removed from request to bidder according to CCPA policy."]
 
             privacy.errors?.isEmpty()
@@ -105,7 +106,7 @@ class CoppaSpec extends PrivacyBaseSpec {
             !privacy.originPrivacy?.ccpa?.usPrivacy
             !privacy.resolvedPrivacy?.ccpa?.usPrivacy
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC].isEmpty()
+            privacy.privacyActionsPerBidder[GENERIC].isEmpty()
 
             privacy.errors?.isEmpty()
         }
@@ -134,7 +135,7 @@ class CoppaSpec extends PrivacyBaseSpec {
             privacy.originPrivacy?.coppa?.coppa == ampStoredRequest.regs.coppa
             privacy.resolvedPrivacy?.coppa?.coppa == ampStoredRequest.regs.coppa
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC] ==
+            privacy.privacyActionsPerBidder[GENERIC] ==
                     ["Geolocation and address were removed from request to bidder according to CCPA policy."]
 
             privacy.errors?.isEmpty()
