+++ src/test/groovy/org/prebid/server/functional/tests/privacy/GdprAmpSpec.groovy
@@ -1,98 +1,34 @@
 package org.prebid.server.functional.tests.privacy
 
 import org.prebid.server.functional.model.ChannelType
-import org.prebid.server.functional.model.bidder.BidderName
+import org.prebid.server.functional.model.config.AccountConfig
+import org.prebid.server.functional.model.config.AccountGdprConfig
+import org.prebid.server.functional.model.config.AccountPrivacyConfig
+import org.prebid.server.functional.model.db.Account
 import org.prebid.server.functional.model.db.StoredRequest
 import org.prebid.server.functional.model.request.auction.BidRequest
-import org.prebid.server.functional.model.response.auction.ErrorType
+import org.prebid.server.functional.testcontainers.PBSTest
 import org.prebid.server.functional.util.privacy.BogusConsent
 import org.prebid.server.functional.util.privacy.CcpaConsent
 import org.prebid.server.functional.util.privacy.TcfConsent
 import spock.lang.PendingFeature
 
+import static org.prebid.server.functional.model.bidder.BidderName.GENERIC
+import static org.prebid.server.functional.model.request.amp.ConsentType.BOGUS
+import static org.prebid.server.functional.model.request.amp.ConsentType.TCF_1
+import static org.prebid.server.functional.model.response.auction.ErrorType.PREBID
+import static org.prebid.server.functional.util.privacy.CcpaConsent.Signal.ENFORCED
 import static org.prebid.server.functional.util.privacy.TcfConsent.GENERIC_VENDOR_ID
 import static org.prebid.server.functional.util.privacy.TcfConsent.PurposeId.BASIC_ADS
 
 @PBSTest
 class GdprAmpSpec extends PrivacyBaseSpec {
 
-
-    def setupSpec(){
+    def setupSpec() {
         cacheVendorList()
     }
-    
-    @PendingFeature
-    def "PBS should add debug log for auction request when valid gdpr was passed"() {
-        given: "Default gdpr BidRequest"
-        def validConsentString = new TcfConsent.Builder()
-                .setPurposesLITransparency(BASIC_ADS)
-                .build()
-
-        def bidRequest = getGdprBidRequest(validConsentString)
-
-        when: "PBS processes auction request"
-        def response = defaultPbsService.sendAuctionRequest(bidRequest)
-
-        then: "Response should contain debug log"
-        assert response.ext?.debug?.privacy
-        def privacy = response.ext?.debug?.privacy
-
-        verifyAll {
-            privacy.originPrivacy?.tcf?.gdpr == "1"
-            privacy.originPrivacy?.tcf?.tcfConsentString == validConsentString as String
-            !privacy.originPrivacy?.tcf?.tcfConsentVersion
-            !privacy.originPrivacy?.tcf?.inEea
-            privacy.resolvedPrivacy?.tcf?.gdpr == "1"
-            privacy.resolvedPrivacy?.tcf?.tcfConsentString == validConsentString as String
-            privacy.resolvedPrivacy?.tcf?.tcfConsentVersion == 2
-            !privacy.resolvedPrivacy?.tcf?.inEea
-
-            !privacy.originPrivacy?.ccpa?.usPrivacy
-            !privacy.resolvedPrivacy?.ccpa?.usPrivacy
-
-            !privacy.originPrivacy?.coppa?.coppa
-            privacy.resolvedPrivacy?.coppa?.coppa == 0
-
-            privacy.privacyActionsPerBidder[BidderName.GENERIC] ==
-                    ["Geolocation was masked in request to bidder according to TCF policy."]
-
-            privacy.errors?.isEmpty()
-        }
-    }
 
     @PendingFeature
-    def "PBS should add debug log for auction request when invalid gdpr was passed"() {
-        given: "Default gdpr BidRequest"
-        def invalidConsentString = new BogusConsent()
-        def bidRequest = getGdprBidRequest(invalidConsentString)
-
-        when: "PBS processes auction request"
-        def response = defaultPbsService.sendAuctionRequest(bidRequest)
-
-        then: "Response should not contain ext.errors"
-        assert !response.ext?.errors
-
-        and: "Response should contain debug log with error"
-        assert response.ext?.debug?.privacy
-        def privacy = response.ext?.debug?.privacy
-
-        verifyAll {
-            privacy.originPrivacy?.tcf?.gdpr == "1"
-            privacy.originPrivacy?.tcf?.tcfConsentString == invalidConsentString as String
-            !privacy.originPrivacy?.tcf?.tcfConsentVersion
-            !privacy.originPrivacy?.tcf?.inEea
-            privacy.resolvedPrivacy?.tcf?.gdpr == "1"
-            privacy.resolvedPrivacy?.tcf?.tcfConsentString == invalidConsentString as String
-            privacy.resolvedPrivacy?.tcf?.tcfConsentVersion == 2
-            !privacy.resolvedPrivacy?.tcf?.inEea
-
-            privacy.privacyActionsPerBidder[BidderName.GENERIC].isEmpty()
-
-            privacy.errors == ["Placeholder: invalid consent string"]
-        }
-    }
-
-     @PendingFeature
     def "PBS should add debug log for amp request when valid gdpr was passed"() {
         given: "AmpRequest with consent string"
         def validConsentString = new TcfConsent.Builder()
@@ -130,7 +66,7 @@ class GdprAmpSpec extends PrivacyBaseSpec {
             !privacy.originPrivacy?.coppa?.coppa
             privacy.resolvedPrivacy?.coppa?.coppa == 0
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC] ==
+            privacy.privacyActionsPerBidder[GENERIC] ==
                     ["Geolocation was masked in request to bidder according to TCF policy."]
 
             privacy.errors?.isEmpty()
@@ -165,7 +101,7 @@ class GdprAmpSpec extends PrivacyBaseSpec {
             privacy.resolvedPrivacy?.tcf?.tcfConsentVersion == 2
             !privacy.resolvedPrivacy?.tcf?.inEea
 
-            privacy.privacyActionsPerBidder[BidderName.GENERIC].isEmpty()
+            privacy.privacyActionsPerBidder[GENERIC].isEmpty()
 
             privacy.errors == ["Amp request parameter consent_string or gdpr_consent have invalid format:" +
                                        " $invalidConsentString" as String]
@@ -187,8 +123,8 @@ class GdprAmpSpec extends PrivacyBaseSpec {
         def response = defaultPbsService.sendAmpRequest(ampRequest)
 
         then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message ==
+        assert response.ext?.errors[PREBID]*.code == [999]
+        assert response.ext?.errors[PREBID]*.message ==
                 ["Amp request parameter gdpr_consent has invalid format for consent type tcfV2: $invalidTcfConsent" as String]
 
         where:
@@ -215,8 +151,8 @@ class GdprAmpSpec extends PrivacyBaseSpec {
         def response = defaultPbsService.sendAmpRequest(ampRequest)
 
         then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message == ["Consent type tcfV1 is no longer supported"]
+        assert response.ext?.errors[PREBID]*.code == [999]
+        assert response.ext?.errors[PREBID]*.message == ["Consent type tcfV1 is no longer supported"]
     }
 
     def "PBS should emit error for amp request with consentString when consent_type is bogus"() {
@@ -239,8 +175,8 @@ class GdprAmpSpec extends PrivacyBaseSpec {
         def response = defaultPbsService.sendAmpRequest(ampRequest)
 
         then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message == ["Invalid consent_type param passed"]
+        assert response.ext?.errors[PREBID]*.code == [999]
+        assert response.ext?.errors[PREBID]*.message == ["Invalid consent_type param passed"]
     }
 
     def "PBS should emit error for amp request when set not appropriate ccpa consent"() {
@@ -261,8 +197,8 @@ class GdprAmpSpec extends PrivacyBaseSpec {
         def response = defaultPbsService.sendAmpRequest(ampRequest)
 
         then: "Response should contain error"
-        assert response.ext?.errors[ErrorType.PREBID]*.code == [999]
-        assert response.ext?.errors[ErrorType.PREBID]*.message ==
+        assert response.ext?.errors[PREBID]*.code == [999]
+        assert response.ext?.errors[PREBID]*.message ==
                 ["Amp request parameter consent_string has invalid format for consent type tcfV2: $ccpaConsent" as String]
     }
 
@@ -326,6 +262,6 @@ class GdprAmpSpec extends PrivacyBaseSpec {
 
         where:
         gdprConfig << [new AccountGdprConfig(enabled: true, channelEnabled: [(ChannelType.AMP): false]),
-                              new AccountGdprConfig(enabled: false)]
+                       new AccountGdprConfig(enabled: false)]
     }
 }
